+++ tests/test_format.py
@@ -27,7 +27,11 @@ paths = [
     os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/analyses/pipelinedamagerepairrate/'),
     os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/epfdamage/'),
     os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/analyses/epfdamage/'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/montecarlofailureprobability/'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/analyses/montecarlofailureprobability/'),
     os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/dataset.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/hazardservice.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/test_hazardservice.py')
 ]