+++ docs/source/conf.py
@@ -33,9 +33,9 @@ project = 'pyIncore'
 author = ''
 
 # The short X.Y version
-version = '0.6.4'
+version = '0.7.0'
 # The full version, including alpha/beta/rc tags
-release = '0.6.4'
+release = '0.7.0'
 
 
 # -- General configuration ---------------------------------------------------