+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added Docker build and release scripts [INCORE1-709](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-709).  
 
 ### Changed
-- Docker file builds documentation using autodoc [INCORE1-648](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-648).
+- Dockerfile builds documentation using autodoc [INCORE1-648](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-648).
 
 ## [0.6.4] - 2020-06-30